DAY 25 - US States Guessing Game
User has to guess 50 us states. After each correct guess, the name of the state is placed on the map. If the user types a secret word 'exit' in the input window, the game creates the list of the states that user has forgotten about (in an additional text file in a csv format). Program is created with the use of PANDAS and TURTLE modules.